+++ app/src/main/java/com/puroblast/weatherappcompose/network/model/Sys.kt
@@ -1,12 +1,6 @@
 package com.puroblast.weatherappcompose.network.model
 
 
-
-
 data class Sys(
-    val country: String,
-    val id: Int,
-    val sunrise: Int,
-    val sunset: Int,
-    val type: Int
+    val country: String, val id: Int, val sunrise: Long, val sunset: Long, val type: Int
 )
